+++ gemfiles/5.0.gemfile
@@ -17,5 +17,4 @@ group :development, :test do
   gem 'selenium-webdriver', '~> 3.0'
   gem 'simplecov', '~> 0.17.1'
   gem 'sqlite3', '~> 1.3.0'
-  gem 'timecop', '~> 0.9.1'
 end
