+++ spec/features/chat_page_spec.rb
@@ -1,9 +1,4 @@
 require 'rails_helper'
-require 'timecop'
-require 'factories/admin_user'
-require 'factories/conversation'
-require 'factories/person'
-require 'factories/text'
 
 feature 'Visit the chat page', js: true do
   given!(:admin) { create(:admin_user) }
@@ -35,7 +30,7 @@ feature 'Visit the chat page', js: true do
   end
 
   scenario 'sends a message' do
-    Timecop.freeze(Time.current) do
+    freeze_time do
       visit admin_chat_path
 
       find('.active-admin-chat__conversation-item', text: person1.email).click
@@ -61,7 +56,7 @@ feature 'Visit the chat page', js: true do
   end
 
   scenario "sends a message and 'No messages' disappear" do
-    Timecop.freeze(Time.current) do
+    freeze_time do
       visit admin_chat_path
 
       find('.active-admin-chat__conversation-item', text: person3.email).click
